Consolidate requests in symbolicate_stacks
The Snappy symbolicator API isn't designed for our use case where
we have many different environments that we want to symbolicate
stacks for. However, we can hack our setting into its format so
we only have to make one request. This change does that, and adds
a few tests for it.